Move functionality to library.
