+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mles"
-version = "2.7.1"
+version = "2.8.0"
 authors = ["jq-rs"]
 edition = "2024"
 description = "Mles, a distributed publish-subscribe data server"
@@ -11,6 +11,14 @@ readme = "README.md"
 categories = ["network-programming"]
 keywords = ["protocol", "publish-subscribe", "websocket", "mles"]
 
+[lib]
+name = "mles"
+path = "src/lib.rs"
+
+[[bin]]
+name = "mles"
+path = "src/main.rs"
+
 [dependencies]
 tokio = { version="1.48", features = ["full"] }
 tokio-stream = { version="0.1.17", features = ["net"] }
@@ -18,8 +26,8 @@ tokio-util = { version="0.7.16", features = ["compat"] }
 tokio-rustls = "0.26"
 rustls-acme = { version="0.14", default-features = false, features = ["tokio", "aws-lc-rs", "webpki-roots"] }
 warp = { version = "0.4", features = ["websocket"] }
-hyper = { version = "1", features = ["server", "http1", "http2"] }
-hyper-util = { version = "0.1", features = ["server", "server-auto", "tokio"] }
+hyper = { version = "1", features = ["server", "http1"] }
+hyper-util = { version = "0.1", features = ["server", "tokio"] }
 simple_logger = "5"
 clap = { version = "4", features = ["derive"] }
 futures-util = { version = "0.3", default-features = false, features = ["sink"] }
@@ -29,6 +37,7 @@ async-compression = { version = "0.4.6", features = ["tokio", "brotli", "zstd"] 
 serde_json = "1.0"
 siphasher = "1.0.0"
 http-types = "2"
+lru = "0.16"
 
 [profile.release]
 lto = true
